Quickfix: Move some classes into correct namespace
